+++ tests/integration/test_read_settings.py
@@ -17,7 +17,7 @@ from vimiv.config import configfile
 #        Configuration dictionaries used to parametrize the configuration file         #
 ########################################################################################
 UPDATED_CONFIG = {
-    "GENERAL": {"shuffle": "True"},
+    "SORT": {"shuffle": "True"},
     "IMAGE": {"overzoom": "4.2"},
     "THUMBNAIL": {"size": "64"},
 }
@@ -44,7 +44,7 @@ UPDATED_EXIF_KEY_SETS = {
 
 
 UPDATED_CONFIG_INVALID = {
-    "GENERAL": {"shuffle": "not a bool"},
+    "SORT": {"shuffle": "not a bool"},
     "IMAGE": {"overzoom": "not a float"},
     "THUMBNAIL": {"size": "not an int"},
 }
@@ -83,7 +83,7 @@ def mock_logger(mocker):
 @pytest.mark.parametrize("configpath", [UPDATED_CONFIG], indirect=["configpath"])
 def test_read_config(configpath):
     """Ensure updated settings are read correctly."""
-    assert api.settings.shuffle.value is True
+    assert api.settings.sort.shuffle.value is True
     assert api.settings.image.overzoom.value == 4.2
     assert api.settings.thumbnail.size.value == 64